pyproject.toml
[tool.poetry]
name = "ws-petproject"
version = "0.1.0"
description = "Backend for developer"
authors = ["matvey tulaev"]


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
celery = "^5.3.5"
redis = "^5.0.1"
djangorestframework = "^3.14.0"
fastapi = "^0.104.1"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
pydantic = "^2.5.3"
isort = "^5.13.2"
pre-commit = "^3.6.0"
gunicorn = "^21.2.0"
djangorestframework-stubs = "^3.14.5"
django-stubs = "^4.2.7"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.12.0"
black = "^23.11.0"
